+++ client/.env.local.example
@@ -9,3 +9,6 @@ GOOGLE_GA_TRACKING_ID =
 NEXT_PUBLIC_SANITY_DATASET=
 NEXT_PUBLIC_SANITY_ID =
 NEXT_SANITY_API_KEY =
+
+#web3 Storage
+NEXT_PUBLIC_WEB3STOTAGE_KEY =
