+++ client/components/index.js
@@ -9,6 +9,5 @@ export { default as SearchBar } from './SearchBar';
 export { default as Filter } from './Filter';
 export { default as NFTCard } from './NFTCard';
 export { default as SenseiProfil } from './SenseiProfil';
-export { default as NFTForm } from './NFTForm';
-export { default as PropertyForm } from './PropertyForm';
+export { default as NFTForm } from './NFTForm/NFTForm';
