+++ client/next.config.js
@@ -4,7 +4,10 @@ const nextConfig = {
 
   compiler:{
     styledComponents: true
-  }
+  },
+  experiments: {
+    topLevelAwait: true,
+ },
 }
 
 module.exports = nextConfig
