+++ client/package.json
@@ -14,19 +14,22 @@
     "ethers": "^5.6.6",
     "googleapis": "^100.0.0",
     "ipfs-http-client": "^56.0.3",
+    "minimist": "^1.2.6",
     "next": "^12.1.6",
     "next-sanity": "^0.5.2",
     "react": "18.1.0",
     "react-dom": "18.1.0",
-    "react-hook-form": "^7.31.1",
+    "react-hook-form": "^7.31.2",
     "react-hyper-modal": "^1.2.16",
     "react-toastify": "^9.0.1",
     "styled-components": "^5.3.5",
     "swiper": "^8.1.4",
     "web3": "^1.7.3",
+    "web3.storage": "^4.2.0",
     "web3modal": "^1.9.7"
   },
   "devDependencies": {
+    "@babel/plugin-syntax-top-level-await": "^7.14.5",
     "eslint": "8.14.0",
     "eslint-config-next": "12.1.6"
   }
